Name Updates
Name officially updates when clicking on a character.

Need to be fixed:

Sliders for Health and Mana are having an issue where they can't turn the stats value into floats.

The Image is not updating, but also not causing errors

Can't seem to click on enemies to make their nameplate appear